\name{readSSAS-package}
\alias{readSSAS-package}
\alias{readSSAS}
\docType{package}
\title{
\packageTitle{readSSAS}
}
\description{
\packageDescription{readSSAS}
}
\details{

Connecting and reading data from Microsoft SQL Server Analysis Service.
Requires configured HTTP Access to Analysis Services on Internet Information Services (IIS)

}
\author{
\packageAuthor{readSSAS}

Maintainer: \packageMaintainer{readSSAS}
}
\keyword{ package }

